fix: compile fails

+++ android_app/app/src/main/java/com/health/openscale/core/bluetooth/scales/HuaweiAH100Handler.kt
@@ -360,7 +360,7 @@ class HuaweiAH100Handler : ScaleDeviceHandler() {
             this.fat = fat
             this.water = water
             if (impedance > 0 && impedance < 4000) {
-                this.impedance = impedance.toFloat()
+                this.impedance = impedance.toDouble()
             }
         }
         publish(m)
